requirements.txt
psutil==5.2.2
PyYAML==3.12
redis==2.10.5
tornado==4.5.1
click==6.7